+++ composer.json
@@ -18,6 +18,11 @@
     "require": {
       "php": ">=7.2"
     },
+    "autoload": {
+      "psr-4": {
+        "theidbrewing\\TT1_Skin\\": "inc/"
+      }
+    },
     "require-dev": {
       "squizlabs/php_codesniffer": "*",
       "dealerdirect/phpcodesniffer-composer-installer": "*",
