+++ tt1skin.php
@@ -19,29 +19,27 @@ if ( ! defined( 'WPINC' ) ) {
 	die;
 }
 
-define( 'TT1SKIN_URL', untrailingslashit( plugin_dir_url( __FILE__ ) ) );
-define( 'TT1SKIN_PATH', untrailingslashit( plugin_dir_path( __FILE__ ) ) );
+define( 'THEIDBREWING_URL', untrailingslashit( plugin_dir_url( __FILE__ ) ) );
+define( 'THEIDBREWING_PATH', untrailingslashit( plugin_dir_path( __FILE__ ) ) );
 
 if ( is_file( 'tt1skin.config.php' ) ) {
 	require_once 'tt1skin.config.php';
 }
 // Json Controller.
-require_once TT1SKIN_PATH . '/inc/class-tt1skin-json-controller.php';
+// require_once THEIDBREWING_PATH . '/inc/class-tt1skin-json-controller.php';
 // TT1 Skin data.
-require_once TT1SKIN_PATH . '/inc/class-tt1skin-skin-data.php';
+// require_once THEIDBREWING_PATH . '/inc/class-tt1skin-skin-data.php';
 
 // TT1 Skin Main Class.
-require_once TT1SKIN_PATH . '/inc/class-tt1skin.php';
-TT1skin::init();
+if ( class_exists( '\theidbrewing\TT1_Skin\Core' ) ) {
+	\theidbrewing\TT1_Skin\Core::init();
+}
 
 // TT1 Skin Block Style.
-require_once TT1SKIN_PATH . '/inc/class-tt1skin-block-style.php';
-TT1skin_Block_Style::init();
+\theidbrewing\TT1_Skin\Block_Style::init();
 
 // TT1 Skin Block Pattern.
-require_once TT1SKIN_PATH . '/inc/class-tt1skin-block-pattern.php';
-TT1skin_Block_Pattern::init();
+\theidbrewing\TT1_Skin\Block_Pattern::init();
 
 // TT1 Skin Customizer Colors.
-require_once TT1SKIN_PATH . '/inc/class-tt1skin-customizer-colors.php';
-TT1skin_Customizer_Colors::register_hooks();
+\theidbrewing\TT1_Skin\Customizer_Colors::register_hooks();
